Test of  multiple line git commit. It took my hours for figure out the bug. Now, enjoy an excerpt from the Popol Vuh creation story:
Here is the story of the beginning,
when there was not one bird,
not one fish,
not one mountain.
Here is the sky, all alone.
Here is the sea, all alone.
There is nothing more
–no sound, no movement.
Only the sky and the sea.
Only Heart-of-Sky, alone.
And these are his names:
Maker and Modeler,
Kukulkan,
and Hurricane.
But there is no one to speak his names.
There is no one to praise his glory.
There is no one to nurture his greatness.